fixed bug; Do not creating a file with a name used
Now, when you try to create a file with a name that
already exists, the konsole will ask if you want to
edit the old file or exit the program to be able to
create a new file with a different name.

The old version of this script, erased the old file
and created a new empty file, losing all the past
work.

+++ Scripts/do.sh
@@ -1,2 +1,11 @@
-cp ~/Documentos/Training-for-Programming-Competitions/'Algorithms, Data Strututes and Others'/Modelo.cpp ./$1.cpp
-vim $1.cpp
+if test -f ./$1.cpp; then
+	echo "$1.cpp already exists!"
+	echo "Do you want to edit it? (y/n)"
+	read option
+	if [ $option == "y" ] ; then
+		vim ./$1.cpp
+	fi
+else
+	cp ~/Documentos/Training-for-Programming-Competitions/'Algorithms, Data Strututes and Others'/Modelo.cpp ./$1.cpp
+	vim $1.cpp
+fi
